+++ src/includes/topics/functions.php
@@ -1479,6 +1479,24 @@ function bbp_split_topic_handler( $action = '' ) {
 						$destination_topic_title = $source_topic->post_title;
 					}
 
+					// Filter the new topic title
+					$destination_topic_title = apply_filters( 'bbp_new_topic_pre_title', $destination_topic_title );
+
+					// Title cannot be empty
+					if ( empty( $destination_topic_title ) ) {
+						bbp_add_error( 'bbp_topic_title', __( '<strong>Error</strong>: Your topic needs a title.', 'bbpress' ) );
+					}
+
+					// Title too long
+					if ( bbp_is_title_too_long( $destination_topic_title ) ) {
+						bbp_add_error( 'bbp_topic_title', __( '<strong>Error</strong>: Your title is too long.', 'bbpress' ) );
+					}
+
+					// Bail before converting the reply if there are errors
+					if ( bbp_has_errors() ) {
+						break;
+					}
+
 					// Update the topic
 					$destination_topic_id = wp_update_post(
 						array(